[Refactor] Moves source code files into src directory
