+++ src/init/InitFormula.h
@@ -5,12 +5,14 @@
 #include "InitData.h"
 #include "../SimInfo.h"
 
-namespace fv2d {
+namespace fv2d
+{
 
-/** 
+/**
  * @brief Abstract class used for initialization of the domain
  */
-struct InitFormula {
+struct InitFormula
+{
   virtual void init(Array Q, const Params &full_params) = 0;
 };
